% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiomassSeries.R
\name{BiomassSeries}
\alias{BiomassSeries}
\title{BiomassSeries Function}
\usage{
BiomassSeries(
  Data,
  param,
  info,
  plot_series = TRUE,
  group,
  grouplabel,
  ylab = "Biomass (1000t)",
  facet = TRUE,
  session
)
}
\arguments{
\item{Data}{RCaNSample_long data-frame computed from the RData RCaNSample}

\item{param}{A character vector containing the species for which the biomass and consumption ratios will be calculated.}

\item{info}{A data frame containing additional species information (e.g., FullName, Colour, Biomass).}

\item{plot_series}{Logical; if TRUE, plots for each series will be created. Default is TRUE.}

\item{group}{Logical; if TRUE, data will be grouped and summarized by year and sample ID. Default is FALSE.}

\item{grouplabel}{A string to be used as the label for the grouped series when `group = TRUE`.}

\item{ylab}{A string specifying the label for the y-axis of the plot. Default is "Biomass (1000t)".}

\item{facet}{Logical; if TRUE, the plot will be faceted for each species. Default is TRUE.}

\item{session}{The Shiny session object, used for adjusting plot text size based on plot width.}
}
\value{
A `ggplot` object that visualizes the biomass data for the selected species, showing quantiles over time.
}
\description{
This function calculates and visualizes biomass data over time for a set of species. It generates a plot
showing quantiles of biomass for each species over the years.
}
