% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsumptionSeries.R
\name{ConsumptionSeries}
\alias{ConsumptionSeries}
\title{ConsumptionSeries Function}
\usage{
ConsumptionSeries(
  Data,
  param,
  info,
  group,
  grouplabel,
  ylab = "Consumption (1000t)",
  facet = FALSE,
  session
)
}
\arguments{
\item{Data}{RCaNSample_long data-frame computed from the RData RCaNSample}

\item{param}{A character vector containing the species for which the biomass and consumption ratios will be calculated.}

\item{info}{A data frame containing additional species information (e.g., FullName, Colour, Biomass).}

\item{group}{Logical; if TRUE, data will be grouped and summarized. Default is FALSE.}

\item{grouplabel}{A string that will be used as the label for the grouped flux series when `group = TRUE`.}

\item{ylab}{A string specifying the label for the y-axis of the plot. Default is "Consumption (1000t)".}

\item{facet}{Logical; if TRUE, the plot will be faceted for each species. Default is FALSE.}

\item{session}{The Shiny session object, used for adjusting plot text size based on plot width.}
}
\value{
A `ggplot` object that visualizes the consumption data for the selected species, with two plots combined
        using `patchwork`.
}
\description{
This function calculates and visualizes consumption fluxes over time for a set of species. It generates
two plots for each species: one for the quantiles of consumption and another showing the proportion
of each prey item.
}
