% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FluxSeries.R
\name{FluxSeries}
\alias{FluxSeries}
\title{FluxSerie Function}
\usage{
FluxSeries(
  Data,
  param,
  info,
  plot_series = TRUE,
  group,
  grouplabel,
  ylab = "Flux (1000t)",
  facet = TRUE,
  session
)
}
\arguments{
\item{Data}{RCaNSample_long data-frame calculated from the RData RCaNSample.}

\item{param}{A vector of strings specifying the fluxes.}

\item{info}{A data frame containing additional information for the species, including their names and colours.}

\item{plot_series}{Logical; whether or not to plot the individual series. Default is TRUE.}

\item{grouplabel}{A string used to label the fluxes (e.g., a custom label for the flux).}

\item{ylab}{A string specifying the label for the y-axis of the plot. Default is "Flux (1000t)".}

\item{facet}{Logical; whether or not to create separate facets for each flux. Default is TRUE.}

\item{session}{The Shiny session object, which is used to get the width of the plot to adjust text size.}
}
\value{
A `ggplot` object that visualizes the quantiles of the flux data.
}
\description{
This function calculates quantiles for fluxes (transitions between prey and predator) over time,
and visualizes them using the `Quantiles_plot` function.
}
