% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fuzzy_proportion_plot.R
\name{Fuzzy_proportion_plot}
\alias{Fuzzy_proportion_plot}
\title{Fuzzy proportion Plot}
\usage{
Fuzzy_proportion_plot(Data, session)
}
\arguments{
\item{Data}{RCaNSample_long data-frame calculated from the RData RCaNSample.}

\item{session}{The Shiny session object, used for responsive plot sizing based on the UI.}

\item{info}{A data frame that contains additional metadata for each target. It should have the columns  `value` ,`series`, `Year`, `target`, and `Colour`.}
}
\value{
A ggplot object representing the proportion bar plot.
}
\description{
Creates a bar plot showing the proportion of the value of each Natural and fishing mortalities with a blurry limit representing the variance of the values.
}
