% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortalitySeries.R
\name{MortalitySeries}
\alias{MortalitySeries}
\title{MortalitySeries Function}
\usage{
MortalitySeries(Data, param, info, group, grouplabel, facet = FALSE, session)
}
\arguments{
\item{Data}{RCaNSample_long data-frame computed from the RData RCaNSample}

\item{param}{A character vector containing the species for which the biomass and consumption ratios will be calculated.}

\item{info}{A data frame containing additional species information (e.g., FullName, Color, Biomass).}

\item{group}{Logical; if TRUE, data will be grouped and summarized. Default is FALSE.}

\item{grouplabel}{A string that will be used as the label for the grouped flux series when `group = TRUE`.}

\item{facet}{Logical; if TRUE, the plot will be faceted for each species. Default is FALSE.}

\item{session}{The Shiny session object, used for adjusting plot text size based on plot width.}
}
\value{
A `ggplot` object that visualizes the predation data for the selected species, with two plots combined
        using `patchwork`.
}
\description{
This function calculates and visualizes mortality over time for a set species. It generates
three plots for each species: one showing the total mortality (Z), one showing the fishing
mortality (F) and one showing the total mortality (G) over time with quantiles.
It is also possible to group the species to compute the mortality of the group.
}
