% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proportion_plot.R
\name{Proportion_plot}
\alias{Proportion_plot}
\title{Proportion Plot}
\usage{
Proportion_plot(Data, info, session)
}
\arguments{
\item{Data}{RCaNSample_long data-frame calculated from the RData RCaNSample.}

\item{info}{A data frame that contains additional metadata for each target. It should have the columns `series`, `FullName`, and `Color`.}

\item{session}{The Shiny session object, used for responsive plot sizing based on the UI.}
}
\value{
A ggplot object representing the proportion bar plot.
}
\description{
Creates a bar plot showing the proportion of the value of each targetted species relative to the total for each year.
This function makes proportion plots for the functions ConsumptionSeries and ProductionSeries.
}
