% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantiles_plot.R
\name{Quantiles_plot}
\alias{Quantiles_plot}
\title{Quantiles Plot with Confidence Intervals and Random trajectories}
\usage{
Quantiles_plot(quantiles, Data, selectedsamples, facet, ylab, session)
}
\arguments{
\item{quantiles}{A data frame containing quantile values (e.g., q0, q2.5, q25, q50, q75, q97.5, q100). Must include columns `Year`, `FullName`, and the quantiles.}

\item{Data}{RCaNSample_long data-frame computed from the RData RCaNSample.}

\item{selectedsamples}{A vector of random trajectories IDs (e.g., 1, 2, 3) that will be used to plot trajectories.}

\item{facet}{A logical value indicating whether the plot should be faceted by `FullName`.}

\item{ylab}{A character string for the label of the y-axis.}

\item{session}{The Shiny session object, passed to retrieve the width of the plot area from the client.}
}
\value{
A ggplot2 object that contains the quantiles plot, including confidence ribbons and sample lines.
}
\description{
This function generates a plot with confidence ribbons for quantiles (0%, 2.5%, 25%, 50%, 75%, 97.5%, 100%)
and sample lines representing selected data. It includes options to facet the plot and customize axis labels.
}
