% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RatioProductionBiomass.R
\name{RatioProductionBiomass}
\alias{RatioProductionBiomass}
\title{Ratio Production to Biomass Plot}
\usage{
RatioProductionBiomass(
  Data,
  param,
  info,
  plot_series = TRUE,
  group,
  grouplabel,
  ylab = "Unitless",
  facet = TRUE,
  session
)
}
\arguments{
\item{Data}{RCaNSample_long data-frame computed from the RData RCaNSample}

\item{param}{A character vector containing the species for which the biomass and production
ratios will be calculated.}

\item{info}{A data frame containing metadata about the species, including columns like `ID`,
`FullName`, `Color`.}

\item{plot_series}{If yes, draw three example of trajectories. Default is `TRUE`.}

\item{group}{A logical value indicating whether the species should be grouped together. Default is `FALSE`.}

\item{grouplabel}{A character string representing the label for the grouped species if `group` is `TRUE`.
Default is `"Default Group Name"`.}

\item{ylab}{A character string for the label of the y-axis. Default is `"Unitless"`.}

\item{facet}{A logical value indicating whether to facet the plot by species. Default is `TRUE`.}

\item{session}{The Shiny session object, passed for client-side session data.}
}
\value{
A ggplot2 object containing the plot of production-to-biomass ratios.
}
\description{
This function calculates the ratio of production to biomass for a set of species and
generates a plot of the range of this ratio over the iterations of RCaN with the example of 3 trajectories.
}
