% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_CaNSample.R
\name{transform_CaNSample}
\alias{transform_CaNSample}
\title{Transform a obj Object into more manipulable Format}
\usage{
transform_CaNSample(obj)
}
\arguments{
\item{obj}{A CaNSample object with `mcmc` results and model structure.}
}
\value{
A tibble with columns: `Sample_id`, `Var`, `Year`, and `value`.
}
\description{
Simplifies a `obj` object by converting the MCMC matrix into a long-format tibble
}
\keyword{internal}
